Data.Version.Package

Description

This module provides functionality for reading a package's version at compile-time.

Since: 0.1.0.0

Type

data PackageVersion

PackageVersion represents PVP version numbers. It is similar to Data.Version's Version except:

PackageVersion has no versionTags.

We enforce PVP's "tags must be at least A.B.C" invariant via the smart-constructor pattern.

Trailing zeroes are ignored in Eq, Ord, Semigroup, and Monoid.

That is, we declare an equivalence class up to trailing zeroes. In particular, the Monoid identity is

[0] = { [0,0,0], [0,0,0,0], ... }

and its Semigroup instance takes the greatest version (based on Ord).

Note: Because we export the underlying list in various ways, (e.g. show), Eq's extensionality can be broken:

x == y ==> f x == f y

Examples

Expand

>>> unsafePackageVersion [0,0,0,0] == unsafePackageVersion [0,0,0]
True

>>> unsafePackageVersion [4,0,0] > unsafePackageVersion [1,2,0,0]
True

>>> unsafePackageVersion [5,6,0] <> unsafePackageVersion [9,0,0]
MkPackageVersion {unPackageVersion = UnsafeRefined {unrefine = [9,0,0]}}

>>> unsafePackageVersion [9,0,0] <> unsafePackageVersion [9,0,0,0]
MkPackageVersion {unPackageVersion = UnsafeRefined {unrefine = [9,0,0]}}

Since: 0.1.0.0

Creation

mkPackageVersion :: [Int] -> Either String PackageVersion

Smart constructor for PackageVersion. The length of the list must be > 2 to match PVP's minimal A.B.C. Furthermore, all digits must be non-negative.

Examples

Expand

>>> mkPackageVersion [1,2,3]
Right (MkPackageVersion {unPackageVersion = UnsafeRefined {unrefine = [1,2,3]}})

>>> mkPackageVersion [2,87,7,1]
Right (MkPackageVersion {unPackageVersion = UnsafeRefined {unrefine = [2,87,7,1]}})

>>> mkPackageVersion [1,2,-3,-4,5]
Left "MkRefineException {predRep = GreaterThanEq 0, targetRep = Int, msg = \"-3 does not satisfy >= 0\"}"

>>> mkPackageVersion [3,2]
Left "MkRefineException {predRep = MinLength 3, targetRep = [Int], msg = \"[3,2] does not satisfy length >= 3\"}"

>>> mkPackageVersion []
Left "MkRefineException {predRep = MinLength 3, targetRep = [Int], msg = \"[] does not satisfy length >= 3\"}"

Since: 0.1.0.0

mkPackageVersionTH :: [Int] -> Q (TExp PackageVersion)

Safely constructs a PackageVersion at compile-time.

Examples

Expand

>>> $$(mkPackageVersionTH [2,4,0])
MkPackageVersion {unPackageVersion = UnsafeRefined {unrefine = [2,4,0]}}

Since: 0.1.0.0

unsafePackageVersion :: [Int] -> PackageVersion

Unsafe version of mkPackageVersion, intended to be used with known constants. Maybe you should use mkPackageVersionTH?

WARNING: This function is not total. Exercise restraint!

Examples

Expand

>>> unsafePackageVersion [1,2,3]
MkPackageVersion {unPackageVersion = UnsafeRefined {unrefine = [1,2,3]}}

Since: 0.1.0.0

fromVersion :: Version -> Either String PackageVersion

Creates a PackageVersion from Version.

Note: Because PackageVersion does not have a versionTags, fromVersion is not injective even on "well-formed" Versions (i.e. non-negative and length > 2). That is, toVersion . fromVersion is not an isomorphism.

Examples

Expand

>>> fromVersion (Version [2,13,0] ["alpha"])
Right (MkPackageVersion {unPackageVersion = UnsafeRefined {unrefine = [2,13,0]}})

>>> fromVersion (Version [] [])
Left "MkRefineException {predRep = MinLength 3, targetRep = [Int], msg = \"[] does not satisfy length >= 3\"}"

Since: 0.1.0.0

fromString :: String -> Either String PackageVersion

Attempts to read a String into a PackageVersion. Leading and/or trailing dots will result in an error, as will the empty string.

Examples

Expand

>>> fromString "1.4.27.3"
Right (MkPackageVersion {unPackageVersion = UnsafeRefined {unrefine = [1,4,27,3]}})

>>> fromString ""
Left "Prelude.read: no parse"

>>> fromString "1.a.2"
Left "Prelude.read: no parse"

>>> fromString ".1.2"
Left "Prelude.read: no parse"

>>> fromString "1.2."
Left "Prelude.read: no parse"

>>> fromString "1.3"
Left "\"MkRefineException {predRep = MinLength 3, targetRep = [Int], msg = \\\"[1,3] does not satisfy length >= 3\\\"}\""

>>> fromString "-3.1.2"
Left "\"MkRefineException {predRep = GreaterThanEq 0, targetRep = Int, msg = \\\"-3 does not satisfy >= 0\\\"}\""

Since: 0.1.0.0

fromText :: Text -> Either String PackageVersion

Attempts to read a Text into a PackageVersion. Leading and/or trailing dots will result in an error, as will the empty string.

Examples

Expand

>>> fromText "1.4.27.3"
Right (MkPackageVersion {unPackageVersion = UnsafeRefined {unrefine = [1,4,27,3]}})

>>> fromText ""
Left "Prelude.read: no parse"

>>> fromText "1.a.2"
Left "Prelude.read: no parse"

>>> fromText ".1.2"
Left "Prelude.read: no parse"

>>> fromText "1.2."
Left "Prelude.read: no parse"

>>> fromText "1.3"
Left "\"MkRefineException {predRep = MinLength 3, targetRep = [Int], msg = \\\"[1,3] does not satisfy length >= 3\\\"}\""

>>> fromText "-3.1.2"
Left "\"MkRefineException {predRep = GreaterThanEq 0, targetRep = Int, msg = \\\"-3 does not satisfy >= 0\\\"}\""

Since: 0.1.0.0

Elimination

toVersion :: PackageVersion -> Version

Creates a Version with empty versionTags from PackageVersion.

Examples

Expand

>>> toVersion (unsafePackageVersion [3,2,0])
Version {versionBranch = [3,2,0], versionTags = []}

Since: 0.1.0.0

toString :: PackageVersion -> String

Displays PackageVersion in String format.

Examples

Expand

>>> toString (unsafePackageVersion [2,7,10,0])
"2.7.10.0"

Since: 0.1.0.0

toText :: PackageVersion -> Text

Displays PackageVersion in Text format.

Examples

Expand

>>> toText (unsafePackageVersion [2,7,10,0])
"2.7.10.0"

Since: 0.1.0.0

Reading Cabal Files

TemplateHaskell

packageVersionTH :: FilePath -> Q (TExp PackageVersion)

TemplateHaskell for reading the cabal file's version at compile-time. Errors encountered will be returned as compilation errors.

Examples

Expand

>>> $$(packageVersionTH "package-version.cabal")
MkPackageVersion {unPackageVersion = UnsafeRefined {unrefine = [0,1,0,0]}}

Since: 0.1.0.0

packageVersionStringTH :: FilePath -> Q (TExp String)

Version of packageVersionTH that returns a string representation of PackageVersion at compile-time. Returns "UNKNOWN" if any errors are encountered.

Examples

Expand

>>> $$(packageVersionStringTH "package-version.cabal")
"0.1.0.0"

>>> $$(packageVersionStringTH "not-found.cabal")
"UNKNOWN"

Since: 0.1.0.0

packageVersionTextTH :: FilePath -> Q (TExp Text)

Version of packageVersionTH that returns a Text representation of PackageVersion at compile-time. Returns "UNKNOWN" if any errors are encountered.

Examples

Expand

>>> $$(packageVersionTextTH "package-version.cabal")
"0.1.0.0"

>>> $$(packageVersionTextTH "not-found.cabal")
"UNKNOWN"

Since: 0.1.0.0

packageVersionEitherTH :: FilePath -> Q (TExp (Either String PackageVersion))

Version of packageVersionTH that returns an Either rather than a compilation error for when something goes wrong.

Examples

Expand

>>> $$(packageVersionEitherTH "package-version.cabal")
Right (MkPackageVersion {unPackageVersion = UnsafeRefined {unrefine = [0,1,0,0]}})

>>> $$(packageVersionEitherTH "not-found.cabal")
Left "not-found.cabal: openFile: does not exist (No such file or directory)"

Since: 0.1.0.0

IO

packageVersionStringIO :: FilePath -> IO String

Version of packageVersionEitherIO that returns a String representation of PackageVersion at runtime. Returns "UNKNOWN" if any errors are encountered.

Examples

Expand

>>> packageVersionStringIO "package-version.cabal"
"0.1.0.0"

>>> packageVersionStringIO "not-found.cabal"
"UNKNOWN"

Since: 0.1.0.0

packageVersionTextIO :: FilePath -> IO Text

Version of packageVersionEitherIO that returns a Text representation of PackageVersion at runtime. Returns "UNKNOWN" if any errors are encountered.

Examples

Expand

>>> packageVersionTextIO "package-version.cabal"
"0.1.0.0"

>>> packageVersionTextIO "not-found.cabal"
"UNKNOWN"

Since: 0.1.0.0

packageVersionEitherIO :: FilePath -> IO (Either String PackageVersion)

Reads the cabal-file's version.

Examples

Expand

>>> packageVersionEitherIO "package-version.cabal"
Right (MkPackageVersion {unPackageVersion = UnsafeRefined {unrefine = [0,1,0,0]}})

>>> packageVersionEitherIO "not-found.cabal"
Left "not-found.cabal: openFile: does not exist (No such file or directory)"

Since: 0.1.0.0