Data.Version.Package

Description

This module provides functionality for reading a package's version at compile-time, along with a type representing PVP version numbers. If only the former is of interest then see packageVersionStringTH, as this is likely the most useful function.

Since: 0.1.0.0

Types

data PackageVersion where

PackageVersion represents PVP version numbers. It is similar to Data.Version's Version (i.e. wraps a [Int]) except:

PackageVersion has no versionTags.

We enforce PVP's "tags must be at least A.B.C" invariant via the smart-constructor pattern.

Trailing zeroes are ignored in Eq, Ord, Semigroup, and Monoid.

That is, we declare an equivalence class up to trailing zeroes. In particular, the Monoid identity is

[0] = { [0,0,0], [0,0,0,0], ... }

and its Semigroup instance takes the greatest version (based on Ord).

Note: Because we export the underlying list in various ways, (e.g. show), Eq's extensionality law,

x == y ==> f x == f y

can be broken. Take care that you do not rely on this law if you are using its underlying [Int] (or String) representation.

Examples

Expand

>>> unsafePackageVersion [0,0,0,0] == unsafePackageVersion [0,0,0]
True

>>> unsafePackageVersion [4,0,0] > unsafePackageVersion [1,2,0,0]
True

>>> unsafePackageVersion [5,6,0] <> unsafePackageVersion [9,0,0]
UnsafePackageVersion {unPackageVersion = [9,0,0]}

>>> unsafePackageVersion [9,0,0] <> unsafePackageVersion [9,0,0,0]
UnsafePackageVersion {unPackageVersion = [9,0,0]}

>>> TR.readEither @PackageVersion "UnsafePackageVersion {unPackageVersion = [3,2,1]}"
Right (UnsafePackageVersion {unPackageVersion = [3,2,1]})

>>> TR.readEither @PackageVersion "UnsafePackageVersion {unPackageVersion = [3,2]}"
Left "Prelude.read: no parse"

Since: 0.1.0.0

Bundled Patterns

pattern MkPackageVersion :: [Int] -> PackageVersion
Since: 0.1.0.0

data ValidationError

Errors that can occur when validating PVP version numbers.

Since: 0.1.0.0

Constructors

VTooShortErr [Int]
PVP version numbers must be at least A.B.C

Since: 0.1.0.0

VNegativeErr Int
PVP version numbers cannot be negative.

Since: 0.1.0.0

data ReadValidateError

Errors that can occur when reading PVP version numbers.

Since: 0.1.0.0

Constructors

RVReadStrErr String
Error when reading a string.

Since: 0.1.0.0

RVValidateErr ValidationError
Validation error.

Since: 0.1.0.0

data ReadFileError

Errors that can occur when reading PVP version numbers from a file.

Since: 0.1.0.0

Constructors

RFFileNotFoundErr String
Error for missing file.

Since: 0.1.0.0

RFVersionNotFoundErr FilePath
Error for missing version.

Since: 0.1.0.0

RFReadValidateErr ReadValidateError
Read/Validation error.

Since: 0.1.0.0

Creation

mkPackageVersion :: [Int] -> Either ValidationError PackageVersion

Smart constructor for PackageVersion. The length of the list must be > 2 to match PVP's minimal A.B.C. Furthermore, all digits must be non-negative.

Examples

Expand

>>> mkPackageVersion [1,2,3]
Right (UnsafePackageVersion {unPackageVersion = [1,2,3]})

>>> mkPackageVersion [2,87,7,1]
Right (UnsafePackageVersion {unPackageVersion = [2,87,7,1]})

>>> mkPackageVersion [1,2,-3,-4,5]
Left (VNegativeErr (-3))

>>> mkPackageVersion [3,2]
Left (VTooShortErr [3,2])

>>> mkPackageVersion []
Left (VTooShortErr [])

Since: 0.1.0.0

mkPackageVersionTH :: [Int] -> Q (TExp PackageVersion)

Safely constructs a PackageVersion at compile-time.

Examples

Expand

>>> $$(mkPackageVersionTH [2,4,0])
UnsafePackageVersion {unPackageVersion = [2,4,0]}

Since: 0.1.0.0

unsafePackageVersion :: [Int] -> PackageVersion

Unsafe version of mkPackageVersion, intended to be used with known constants. Maybe you should use mkPackageVersionTH?

WARNING: This function is not total. Exercise restraint!

Examples

Expand

>>> unsafePackageVersion [1,2,3]
UnsafePackageVersion {unPackageVersion = [1,2,3]}

>>> unsafePackageVersion [1,2]
UnsafePackageVersion {unPackageVersion = *** Exception: PVP numbers must be at least A.B.C: [1, 2]
CallStack (from HasCallStack):
  error, called at src/Data/Version/Package.hs:362:32 in main:Data.Version.Package

Since: 0.1.0.0

fromVersion :: Version -> Either ValidationError PackageVersion

Creates a PackageVersion from Version.

Note: Because PackageVersion does not have a versionTags, fromVersion is not injective even on "well-formed" Versions (i.e. non-negative and length > 2). That is, toVersion . fromVersion is not an isomorphism.

Examples

Expand

>>> fromVersion (Version [2,13,0] ["alpha"])
Right (UnsafePackageVersion {unPackageVersion = [2,13,0]})

>>> fromVersion (Version [] [])
Left (VTooShortErr [])

Since: 0.1.0.0

fromString :: String -> Either ReadValidateError PackageVersion

Attempts to read a String into a PackageVersion. Leading and/or trailing dots will result in an error, as will the empty string.

Examples

Expand

>>> fromString "1.4.27.3"
Right (UnsafePackageVersion {unPackageVersion = [1,4,27,3]})

>>> fromString ""
Left (RVReadStrErr "Prelude.read: no parse")

>>> fromString "1.a.2"
Left (RVReadStrErr "Prelude.read: no parse")

>>> fromString ".1.2"
Left (RVReadStrErr "Prelude.read: no parse")

>>> fromString "1.2."
Left (RVReadStrErr "Prelude.read: no parse")

>>> fromString "1.3"
Left (RVValidateErr (VTooShortErr [1,3]))

>>> fromString "-3.1.2"
Left (RVValidateErr (VNegativeErr (-3)))

Since: 0.1.0.0

fromText :: Text -> Either ReadValidateError PackageVersion

Attempts to read a Text into a PackageVersion. Leading and/or trailing dots will result in an error, as will the empty string.

Examples

Expand

>>> fromText "1.4.27.3"
Right (UnsafePackageVersion {unPackageVersion = [1,4,27,3]})

>>> fromText ""
Left (RVReadStrErr "Prelude.read: no parse")

>>> fromText "1.a.2"
Left (RVReadStrErr "Prelude.read: no parse")

>>> fromText ".1.2"
Left (RVReadStrErr "Prelude.read: no parse")

>>> fromText "1.2."
Left (RVReadStrErr "Prelude.read: no parse")

>>> fromText "1.3"
Left (RVValidateErr (VTooShortErr [1,3]))

>>> fromText "-3.1.2"
Left (RVValidateErr (VNegativeErr (-3)))

Since: 0.1.0.0

Elimination

unPackageVersion :: PackageVersion -> [Int]

Since: 0.1.0.0

toVersion :: PackageVersion -> Version

Creates a Version with empty versionTags from PackageVersion.

Examples

Expand

>>> toVersion (UnsafePackageVersion [3,2,0])
Version {versionBranch = [3,2,0], versionTags = []}

Since: 0.1.0.0

toString :: PackageVersion -> String

Displays PackageVersion in String format.

Examples

Expand

>>> toString (UnsafePackageVersion [2,7,10,0])
"2.7.10.0"

Since: 0.1.0.0

toText :: PackageVersion -> Text

Displays PackageVersion in Text format.

Examples

Expand

>>> toText (UnsafePackageVersion [2,7,10,0])
"2.7.10.0"

Since: 0.1.0.0

Reading Cabal Files

TemplateHaskell

These functions allow for reading a cabal's version at compile-time. If the intention is to simply read the value so it can be printed during runtime (e.g. for an executable's --version flag), then packageVersionStringTH (or packageVersionTextTH) is the best choice, as any errors encountered will not prevent compilation.

packageVersionTH :: FilePath -> Q (TExp PackageVersion)

TemplateHaskell for reading the cabal file's version at compile-time. Errors encountered will be returned as compilation errors.

Examples

Expand

>>> $$(packageVersionTH "package-version.cabal")
UnsafePackageVersion {unPackageVersion = [0,1,0,0]}

Since: 0.1.0.0

packageVersionStringTH :: FilePath -> Q (TExp String)

Version of packageVersionTH that returns a String representation of PackageVersion at compile-time. Returns "UNKNOWN" if any errors are encountered.

Examples

Expand

>>> $$(packageVersionStringTH "package-version.cabal")
"0.1.0.0"

>>> $$(packageVersionStringTH "not-found.cabal")
"UNKNOWN"

Since: 0.1.0.0

packageVersionTextTH :: FilePath -> Q (TExp Text)

Version of packageVersionTH that returns a Text representation of PackageVersion at compile-time. Returns "UNKNOWN" if any errors are encountered.

Examples

Expand

>>> $$(packageVersionTextTH "package-version.cabal")
"0.1.0.0"

>>> $$(packageVersionTextTH "not-found.cabal")
"UNKNOWN"

Since: 0.1.0.0

IO

packageVersionThrowIO :: FilePath -> IO PackageVersion

Version of packageVersionEitherIO that throws an Exception if any errors are encountered.

Examples

Expand

>>> packageVersionThrowIO "package-version.cabal"
UnsafePackageVersion {unPackageVersion = [0,1,0,0]}

Since: 0.1.0.0

packageVersionStringIO :: FilePath -> IO String

Version of packageVersionEitherIO that returns a String representation of PackageVersion at runtime. Returns "UNKNOWN" if any errors are encountered.

Examples

Expand

>>> packageVersionStringIO "package-version.cabal"
"0.1.0.0"

>>> packageVersionStringIO "not-found.cabal"
"UNKNOWN"

Since: 0.1.0.0

packageVersionTextIO :: FilePath -> IO Text

Version of packageVersionEitherIO that returns a Text representation of PackageVersion at runtime. Returns "UNKNOWN" if any errors are encountered.

Examples

Expand

>>> packageVersionTextIO "package-version.cabal"
"0.1.0.0"

>>> packageVersionTextIO "not-found.cabal"
"UNKNOWN"

Since: 0.1.0.0

packageVersionEitherIO :: FilePath -> IO (Either ReadFileError PackageVersion)

Reads the cabal-file's version.

Examples

Expand

>>> packageVersionEitherIO "package-version.cabal"
Right (UnsafePackageVersion {unPackageVersion = [0,1,0,0]})

Since: 0.1.0.0